+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-10-19 03:07:36
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 7c89c678360f9d87384f8e3e388cb8f25ede5c36
+
+## [ForwardWidgets-2kuai] - 2025-10-19 03:07:36
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 589f7ff04ebfffe8385ead46d9a8115af70b56a3
+
+## [ForwardWidgets-pack1r] - 2025-10-19 03:07:36
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: dcf6532ba6a11c831fca71f4c2077cb2c22824c5
+
+
 ## [ForwardWidgets-huangxd] - 2025-10-18 02:45:22
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main